Check outcome of previously issued reboots.
With this change Rebot will save a Status field in the history JSON. Such Status
can be Unchecked, Rebooted or Failed. An issued reboot command saves the node
as "Unchecked". On the following run, all the Unchecked nodes will become
either Failed or Rebooted, depending on whether they are still detected as to
be rebooted or not.